requirements.txt
numpy>=1.20.0
Pillow>=9.0.0
# Vercel serverless 自动安装以上依赖
